// src/types/place.ts
export type Place = {
  id: number;
  slug: string;
  name: string;
  continent: string;
  country: string;
  region: string;
  city: string;
  type: string;
  gf_category: string;
  address: string;
  latitude: number;
  longitude: number;
  phone: string;
  website: string;
  description: string;
  notes: string;
  is_demo: boolean;
  verified: boolean;
  published: boolean;
  delivery_available: boolean;
  direct_delivery: boolean;
  takeaway_available: boolean;
  just_eat_url: string;
  glovo_url: string;
  too_good_to_go_available: boolean;
  too_good_to_go_url: string;
  created_at?: string;
  updated_at?: string;
};